+++ packages/Button/CHANGELOG.md
@@ -1,7 +1,7 @@
-# @mindfului/button.react
+# @ladder-ui/button.react
 
 ## 0.1.1
 
 ### Patch Changes
 
-- 98fa1ec: Initial experimental version for mindful ui pacakges
+- 98fa1ec: Initial experimental version for ladder ui pacakges